Rename pipeline_tasks terminating_jobs.py to jobs_terminating.py
